requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
ffmpeg-python==0.2.0
tenacity==8.3.0
openai==1.30.1
Pillow==10.3.0
